fix: read item catalog bigint identifiers without narrowing

+++ src/God2.ServerV2.Persistence/MariaDbItemStackRuleRepository.cs
@@ -44,7 +44,7 @@ public sealed class MariaDbItemStackRuleRepository(
                 $"Invalid maximum_stack: item={itemId}");
 
         return new ItemStackRule(
-            reader.GetInt32(0),
+            reader.GetInt64(0),
             reader.GetString(1),
             configuredMaximum,
             reader.GetBoolean(3));
